requirements.txt
# Core dependencies
python-telegram-bot==20.7
motor==3.3.2
python-dotenv==1.0.0
pymongo==4.6.1

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Redis for caching (optional)
redis==5.0.1
aioredis==2.0.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Image processing (for player cards)
Pillow==10.1.0

# Logging and monitoring
colorlog==6.8.0
colorama==0.4.6
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Timezone support
pytz==2023.3.post1

# Security
cryptography==41.0.7

# Performance monitoring
psutil==5.9.6

# API documentation
pydantic==2.5.3